[IMP] helpdesk: improve the helpdesk overview
Purpose of the commit is to improve the helpdesk overview.

So in this commit did the following changes:
Generic:
- added an 'open tickets' label under 'my tickets'
- renamed 'sla failed' into 'failed tickets'
- renamed 'avg 7 days' into 'avg last 7 days'
- display 'open tickets' in green if > 0
- display failed tickets in orange if > 0
- display avg open hours in grey
- moved the 'happy rating' column to the right of 'success rate'
- clicking anywhere on the daily target cell should switch it to edit mode
- renamed various action name

Tickets:
- the avg open hours point to the helpdesk.ticket.report.analysis
  model instead of helpdesk.ticket
- the default value for pivot view of avg open hours:
  group by: create date > day
  measure: hours open, count
  filter: my tickets and open tickets
- the default value for graph view of avg open hours:
  group by: create date > day
  measure: hours open
  filter: my tickets and open tickets
  in barchart view, unstacked and in descending order
- added cohort view in the avg open hours
- the default value for cohort view of avg open hours:
  mode: day
  measure: hours open
- the default value for closed tickets > today:
  open a list view with a default filter on 'my tickets' and 'closed tickets'
  point to the helpdesk.ticket instead of the helpdesk.ticket.report.analysis
- the default value for closed tickets > avg 7 days:
  group by: close date > day
  filter: my tickets and close tickets

Ratings
- renamed 'happy rating' into 'customer satisfaction'
- default filter: my ratings
- the default value for pivot view of today:
  group by: subject
- the default value for graph view of today:
  group by: document
- the default value for graph view of avg 7 days:
  group by: date > day

Success rate
- today: open a list view by default
  default filter: closed tickets, my tickets, sla success
- avg 7 days:
 default filter: my tickets, sla success and closed tickets
default group by: closed date > day

task-2531487
closes #19809

Related: odoo/upgrade#3119

+++ helpdesk/models/res_users.py
@@ -7,7 +7,7 @@ from odoo import fields, models
 class ResUsers(models.Model):
     _inherit = 'res.users'
 
-    helpdesk_target_closed = fields.Float(string='Target Tickets to Close', default=1)
+    helpdesk_target_closed = fields.Integer(string='Target Tickets to Close', default=1)
     helpdesk_target_rating = fields.Float(string='Target Customer Rating', default=100)
     helpdesk_target_success = fields.Float(string='Target Success Rate', default=100)
